/home/repa/haskell/bayes_classifier/bayesClassifier.hs autocommit from emacs

+++ bayes_classifier/bayesClassifier.hs
@@ -62,7 +62,7 @@ main = do
       sums = fromListVector (Z :. nLabel) $ map sumAllS likelyhoods
       
       [trainData'] = computeP $ traverse likelyhoods' (\(Z :. i :. j) -> (Z :. i :. j))
-                                                      (\lp (Z :. i :. j) -> ((lp (Z :. i :. j))) / 
+                                                      (\lp (Z :. i :. j) -> (fromIntegral (lp (Z :. i :. j)):: Double) / 
                                                                              fromIntegral (sums ! (Z :. i)))
 
 --      likelyhoods is append after I sum across all indexes -> Array U (Z :. labelIdx) (Array U (Z :. WordIdx) Int)
